start of tests for refactor.. Lots to do.. need doing urgently and normalk service resuming
